Client and the Two Servers
I implemented all of the steps for what the things are supposed to do
upon recieving a certain type of signal from the sockets and such;
however I did not write the code to listen to the sockets and get it to
run that function, nor did I write the code to actually send the data
over the sockets. As I told you, the code for recieving data/sending
data over the sockets, when it knows that it needs to be done, should be
written in sendData and recieveData. I am assuming that the sockets that
the servers are going to be on are known, and are of the form CLA_SOCKET
and CTF_SOCKET. Other than that, everything in thee code should be
syntactically correct unless I forgot some things with my python.